requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
python-dotenv==1.0.0
SQLAlchemy==2.0.20
Werkzeug==2.3.7
WTForms==3.0.1
email-validator==2.0.0
python-dateutil==2.8.2
requests==2.31.0
Pillow==10.0.0
PyPDF2==3.0.1
alembic==1.12.0
reportlab
# Keycloak integration dependencies
authlib==1.3.0
python-keycloak==3.7.0
cryptography
PyJWT==2.8.0
